+++ src/RoadRegistry.BackOffice/Messages/RoadNetworkExtractCloseReason.cs
@@ -2,7 +2,6 @@ namespace RoadRegistry.BackOffice.Messages;
 
 public enum RoadNetworkExtractCloseReason
 {
-    SuccessfullUploadReceived,
     NoDownloadReceived,
     InformativeExtract
 }