+++ src/RoadRegistry.Editor.Projections/ExtractRequestRecordProjection.cs
@@ -6,10 +6,13 @@ using Be.Vlaanderen.Basisregisters.ProjectionHandling.Connector;
 using Be.Vlaanderen.Basisregisters.ProjectionHandling.SqlStreamStore;
 using Microsoft.EntityFrameworkCore;
 using NetTopologySuite.Geometries;
+using NodaTime.Text;
 using Schema;
 using Schema.Extracts;
 using System;
 using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 
 public class ExtractRequestRecordProjection : ConnectedProjection<EditorContext>
 {
@@ -24,7 +27,7 @@ public class ExtractRequestRecordProjection : ConnectedProjection<EditorContext>
                 ExternalRequestId = message.ExternalRequestId,
                 Contour = (Geometry)GeometryTranslator.Translate(message.Contour),
                 DownloadId = message.DownloadId,
-                Description = message.Description,
+                Description = !string.IsNullOrEmpty(message.Description) ? message.Description : "onbekend",
                 IsInformative = message.IsInformative
             };
             await context.ExtractRequests.AddAsync(record, ct);
@@ -39,17 +42,52 @@ public class ExtractRequestRecordProjection : ConnectedProjection<EditorContext>
                 ExternalRequestId = message.ExternalRequestId,
                 Contour = (Geometry)GeometryTranslator.Translate(message.Contour),
                 DownloadId = message.DownloadId,
-                Description = message.Description,
+                Description = !string.IsNullOrEmpty(message.Description) ? message.Description : "onbekend",
                 IsInformative = message.IsInformative
             };
             await context.ExtractRequests.AddAsync(record, ct);
         });
 
+        When<Envelope<RoadNetworkExtractDownloaded>>(async (context, envelope, ct) =>
+        {
+            var record = context.ExtractRequests.Local.SingleOrDefault(record => record.DownloadId == envelope.Message.DownloadId)
+                                 ?? await context.ExtractRequests.SingleAsync(record => record.DownloadId == envelope.Message.DownloadId, ct);
+
+            record.DownloadedOn = InstantPattern.ExtendedIso.Parse(envelope.Message.When).Value.ToDateTimeOffset();
+        });
+
         When<Envelope<RoadNetworkExtractClosed>>(async (context, envelope, ct) =>
         {
-            var record = context.ExtractRequests.Local.SingleOrDefault(record => record.ExternalRequestId == envelope.Message.ExternalRequestId)
-                ?? await context.ExtractRequests.SingleAsync(record => record.ExternalRequestId == envelope.Message.ExternalRequestId, ct);
-            record.IsInformative = true;
+            await CloseExtractRequests(context, envelope.Message.ExternalRequestId, ct);
         });
+
+        When<Envelope<RoadNetworkChangesAccepted>>(async (context, envelope, ct) =>
+        {
+            if (envelope.Message.DownloadId is null)
+            {
+                return;
+            }
+            
+            var downloadRecord = context.ExtractRequests.Local.SingleOrDefault(record => record.DownloadId == envelope.Message.DownloadId.Value)
+                ?? await context.ExtractRequests.SingleAsync(record => record.DownloadId == envelope.Message.DownloadId.Value, ct);
+
+            await CloseExtractRequests(context, downloadRecord.ExternalRequestId, ct);
+        });
+    }
+
+    private async Task CloseExtractRequests(EditorContext editorContext, string externalRequestId, CancellationToken cancellationToken)
+    {
+        var records = editorContext.ExtractRequests.Local
+            .Where(record => record.ExternalRequestId == externalRequestId)
+            .ToList()
+            .Concat(await editorContext.ExtractRequests
+                .Where(record => record.ExternalRequestId == externalRequestId)
+                .ToListAsync(cancellationToken))
+            ;
+
+        foreach (var record in records)
+        {
+            record.IsInformative = true;
+        }
     }
 }