+++ src/RoadRegistry.Editor.Schema/Migrations/20230606085359_RemoveUnusedRoadNodeAndRoadSegmentBoundingBoxTables.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using NetTopologySuite.Geometries;
 using RoadRegistry.Editor.Schema;
@@ -12,9 +13,10 @@ using RoadRegistry.Editor.Schema;
 namespace RoadRegistry.Editor.Schema.Migrations
 {
     [DbContext(typeof(EditorContext))]
-    partial class EditorContextModelSnapshot : ModelSnapshot
+    [Migration("20230606085359_RemoveUnusedRoadNodeAndRoadSegmentBoundingBoxTables")]
+    partial class RemoveUnusedRoadNodeAndRoadSegmentBoundingBoxTables
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -292,9 +294,6 @@ namespace RoadRegistry.Editor.Schema.Migrations
                     b.Property<long>("AvailableOn")
                         .HasColumnType("bigint");
 
-                    b.Property<DateTimeOffset?>("DownloadedOn")
-                        .HasColumnType("datetimeoffset");
-
                     b.Property<string>("ExternalRequestId")
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
@@ -328,9 +327,6 @@ namespace RoadRegistry.Editor.Schema.Migrations
                     b.Property<string>("Description")
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
-
-                    b.Property<DateTimeOffset?>("DownloadedOn")
-                        .HasColumnType("datetimeoffset");
 
                     b.Property<string>("ExternalRequestId")
                         .IsRequired()