+++ src/RoadRegistry.Wfs.Projections/RoadSegmentRecordProjection.cs
@@ -2,14 +2,12 @@ namespace RoadRegistry.Wfs.Projections;
 
 using System;
 using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using BackOffice;
 using BackOffice.Messages;
 using Be.Vlaanderen.Basisregisters.ProjectionHandling.Connector;
 using Be.Vlaanderen.Basisregisters.ProjectionHandling.SqlStreamStore;
-using Be.Vlaanderen.Basisregisters.Shaperon;
 using Microsoft.EntityFrameworkCore;
 using Schema;
 using Syndication.Schema;
@@ -179,7 +177,10 @@ public class RoadSegmentRecordProjection : ConnectedProjection<WfsContext>
 
         var roadSegmentRecord = await context.RoadSegments.FindAsync(roadSegmentModified.Id).ConfigureAwait(false);
 
-        if (roadSegmentRecord == null) throw new InvalidOperationException($"RoadSegmentRecord with id {roadSegmentModified.Id} is not found");
+        if (roadSegmentRecord == null)
+        {
+            throw new InvalidOperationException($"RoadSegmentRecord with id {roadSegmentModified.Id} is not found");
+        }
 
         roadSegmentRecord.Id = roadSegmentModified.Id;
         roadSegmentRecord.BeginTime = LocalDateTimeTranslator.TranslateFromWhen(envelope.Message.When);
@@ -275,7 +276,11 @@ public class RoadSegmentRecordProjection : ConnectedProjection<WfsContext>
     private static async Task RemoveRoadSegment(RoadSegmentRemoved roadSegmentRemoved, WfsContext context)
     {
         var roadSegmentRecord = await context.RoadSegments.FindAsync(roadSegmentRemoved.Id).ConfigureAwait(false);
-        if (roadSegmentRecord == null) return;
+        if (roadSegmentRecord == null)
+        {
+            throw new InvalidOperationException($"RoadSegmentRecord with id {roadSegmentRemoved.Id} is not found");
+        }
+
         context.RoadSegments.Remove(roadSegmentRecord);
     }